+++ RELEASES.rst
@@ -1,5 +1,6 @@
 **Version:** 1.0.1
 - Add owner on FileUploaderHistory
+- Rename FileUploaderHistory.content to FileUploaderHistory.filename
 
 **Version:** 1.0.0
 - Start project